fastapi-be/pyproject.toml
[project]
name = "fastapi-be"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "baml-py (>=0.213.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "fastapi (>=0.121.1,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "openai (>=2.7.2,<3.0.0)",
    "e2b-code-interpreter (>=2.3.0,<3.0.0)",
    "e2b (>=2.6.4,<3.0.0)",
    "deepagents (>=0.2.7,<0.3.0)",
    "langchain (>=1.0.7,<2.0.0)",
    "langchain-core>=1.0.5,<2.0.0",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "tavily-python>=0.7.13",
]
